Change generated voicemail greetings to a shorter prompt set.
Added the prompt owner.wav, which plays this:
   "The owner of this extension"

Changed the generated voicemail greetings to use that prompt instead
of voicing the extension characters one by one if their are non-digits
in the extension, or if the extension is greater than 5 characters long.

This only takes effect if the user selects something OTHER than the
"default greeting", which uses VoiceXML to speak all the digits, instead
of using the generated greetings.  I think this is not correct, but I'll
leave it for now.

Fixes XECS-14.

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@10070 ab1d8caa-1f67-47f1-9e81-24633a41865c