Patch to play/record RIFF files on big endian systems
Provided by Misha Vodsedalek.  He writes:

I noticed that Scott submitted a fix for XECS-24. With this fix, the media
server should work fine on big endian platforms provided wave files are
converted to the RIFX format (from RIFF). This is fine as far as we stay on
the big endian platform. 
  
The problem with this approach is that Windows applications (such as the
Windows Media Player) do not understand the RIFX format. When sipXecs sends
the wave file in the RIFX format as an e-mail attachment or when voice mail is
accessed via the Web interface, attempts to play the message on a little
endian (X86) machine unfortunately fail. Well - I failed to play the RIFX file
on my Windows XP. Maybe Vista can handle it, but I wouldn't bet on it. 
  
Anyway, I used another approach - my code changes in the media server allow
sipXecs to use RIFF files on the big endian platform. Due to that, there is no
conversion required when building/installing sipXecs and the compatibility
with X86 applications that want to listen to sipXecs generated voice messages
is ensured. I could make a few minor enhancements to allow sipXecs to read
both RIFF and RIFX, but for the output I'd use just RIFF (to guarantee the
above mentioned compatibility with Windows).

...

This patch allows RIFF files to be used on big endian systems. Based on the
feeback from Pingtel designers, support for RIFX has not been implemented.
Changes to use RIFX files under HP Unix have been removed.


Resolves XECS-24 (differently than the previous resolution)

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@10180 ab1d8caa-1f67-47f1-9e81-24633a41865c