Try harder to shutdown processes--complain louder if they won't.
After sending SIGTERM, it now waits up to 20 seconds for the process to exit.

After that, it sends SIGABRT (in an attempt to get a core dump), and waits
up to 60 seconds for the process to exit.

After that, it sends SIGKILL, and waits up to 60 more seconds.

If the darn thing is still alive after all that, a warning is displayed.

Here is what it looks like:
$ /etc/init.d/sipxpbx stop
Stopping sipXpbx: 
  Stopping: watchdog 
  Stopping: keepalive 
...
  Confirm Stop: watchdog                                   [  OK  ]
  Confirm Stop: keepalive                                  [  OK  ]
...
  Confirm Stop: sipxacd ....................
    sipxacd did not Stop.
    Trying to abort sipxacd  55 50 45 40 35 30 25 20 15 10 5 0
    sipxacd did not abort.
    Trying to kill sipxacd  55 50 45 40 35 30 25 20 15 10 5 0
                                                           [FAILED]
    sipxacd will not die.
    Machine must be restarted!
  Confirm Stop: sipxcallresolver                           [  OK  ]
...

The "OK" is in the color green on capable terminals, while the "FAILED" and 
restart message are in the color red.

Fixes XECS-113 - Shutting down processes that do not shut down promptly has 
several problems (esp. no core dump taken)

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@10250 ab1d8caa-1f67-47f1-9e81-24633a41865c