Make fastdb use original semaphore key generation algorithms
The semaphore cleanup code in ipcclean.pl relies on the fastdb using
the standard ftok() method of generating a key based on the inode of
a file.

The new fastdb code used a different method, and as such the semaphores
were not getting cleaned up correctly as fastdb and ipcclean didn't 
agree on the keys.  This was leaving the semaphores behind, but deleting
the files.  

Startup problems can occur if the files weren't created with the exact
same inode numbers they had before, causing semaphore key mismatches,
causing the imdb to fail.

Changed the new fastdb code to use the original method.

Fixes XECS-547 - imdb fails to initialize on third startup

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11175 ab1d8caa-1f67-47f1-9e81-24633a41865c