XCF-1965 Intercom feature should be extended to support other phone vendors
Enable sipXconfig to generate a mappingrules.xml file that works with LG, Snom,
and possibly other phones that support the Call-Info header for controlling auto-answer.

Internationalization of the intercom_*.properties files is also required, as the English
version of intercom.properties has changed to include LG, and Snom phones into the list of supported devices.

Patch from woof! - thanks!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11189 ab1d8caa-1f67-47f1-9e81-24633a41865c