Check return value of getReplacesData to see if it found anything.
Newer version of getReplacesData zeros out data structures if nothing was
found, previous version returned what was passed in.

It is possible the now "empty" callId field was being compared with
other calls and matching, causing the messages to be directed to the
incorrect CpCall thread, creating havoc.  By checking the return value
of getReplacesData, this is avoided.  In any event, it's a good thing
to do.

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11270 ab1d8caa-1f67-47f1-9e81-24633a41865c