XCF-1958 make sure that firt-run task is not executed during DB upgrade phase
Usually initialization tasks are executed during upgrade phase of sipXconfig startup.
It is too early for first-run task to be executed then (it requires that replication threads are running and functional)
I made FirstRunTask to execute only when it receives application initialization even during 'normal' phase.
There are at least 2 hacks here:
- need to make sure that DataInitializer does not remove first-run tasks
- AdminContext now provides inUpgradePhase method that can be used to determine which phase we are in. The implementation is less than idead (checks for 'tapestry' bean which is only present in normal phase)

I suspect that we will be adjusting this implementation shortly when the watchdog is reworked not to start other sipx services in first-run phase.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11330 ab1d8caa-1f67-47f1-9e81-24633a41865c