Add credentials to ACD lines to make transfers work.
This change also modifies the way ACDLines are constructed:
Previously, the line identity was:
  queue-name@acd-host:port
with a sipXtapi alias for that defined as:
  extension@domain

The extension@domain form of the address is configured as
an alias in the redirect server (by sipXconfig) to point to
the queue-name@acd-host:port.

The sipXtapi lines (and therefor, the lower level SipLine objects)
are now defined using the extension@domain form as the line
identifier.

This change has been made in ACDLineManager - the configuration file
has not been changed, just how the fields it provides are used.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11533 ab1d8caa-1f67-47f1-9e81-24633a41865c