XCF-2059 fix editing auto refresh interval
At some point refresh interval became parameter of AutoRefreshFrom which made it read-only.
Introduced a property that is initialized from this parameter: value of the property can be safely changed.
We need to upgrade UI test suite so that we do not break it again: current UI tests do not deal well with Ajax enabled components.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11590 ab1d8caa-1f67-47f1-9e81-24633a41865c