Don't bother to set display name on ACD Lines.
Name being used was actually the Description field fom the WebUI, 
which is free form and not checked to be a valid display name and 
setDisplayName will abort if given a bad name.

As the outbound calls set the name to be that of the inbound caller,
I don't see any reason to set this to anything.

Fixes XECS-1197 - acd not running on sipxchange.pingtel.com since 3.10 upgrade

--Woof!



git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@12056 ab1d8caa-1f67-47f1-9e81-24633a41865c