XCF-2095 LDAP import: trim 'namingContext' value to support servers returning strange values.
The Domino server is returning a null byte for the namingContexts value.
In Java, strings are not null-terminated and the 00 is taken as an actual string (displayed in the logs as ^@). Trying to save this to DB causes Postgres exceptions.

The attached patch trims the string received from the LDAP server to remove white space, including null bytes, and only saves it as searchBase if it is not empty.

Patch from Carolyn Beeton - thanks!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@12065 ab1d8caa-1f67-47f1-9e81-24633a41865c