XCF-2499 fix LdapServer page: displaying page and displaying errors on the page
The direct cause for the problem was a missing space in the page specification.  Added a display
test (most of the page depends on LDAP server running but we can at least get to the first error).

Also fixed displaying non-localized message errors.  UserException has now a special constructor
that allows for passing a property key rather than hardcoded message. If key is specified
UserExceptionFilter retrieves the message from property bundle associated with the page on which
exception is caught. This is still not ideal, but should work for a few localized error messages
that we have.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@12493 ab1d8caa-1f67-47f1-9e81-24633a41865c