sipXpage: Send ACK and BYE to forked 200 OKs from page INVITEs.
For the 2nd and subsequent 200 OKs from an INVITE, send the ACK out 
of dialog, so as not to confuse JAIN-SIP that we are accepting this 
new dialog as the main one (it can only handle one at at time).  Also 
create a new transaction and send a BYE so as not to leave the answered
call stranded.

This only occurs in the case of a forked INVITE that multiple devices 
answer simultainously.  The proxy will select the on that answers first, 
and send CANCELs to all the others, but it is possible the phones answer 
before the CANCEL, and so the additional 200 OKs are forwarded by the proxy 
to sipXpage.  JAIN-SIP cannot handle multiple dialogs, and so one (the 
first on it sees) is selected, and the others are dropped.

It is still not possible to page multiple devices via proxy forking.  
The first one to answer "wins", all others are dropped.

Requires on JAIN-SIP rev 1.2.84 or later in sipXcommons.

Fixes XECS-1591 - Paging server mishandles page to a user with multiple phones

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@13607 ab1d8caa-1f67-47f1-9e81-24633a41865c