XCF-2819: Added the "Route" field back to the initial screen for SIP trunk creation.
To avoid the earlier problems associated with making the gateway being
edited persistent (see XCF-2444), a transient (that is, not persisted to
the DB) is used for uncommitted changes. That transient instance is made
persistent when leaving to go to the "Create SBC" page.

The persisted Gateway instance is cleared out after returning from the
"Create SBC" page, or when navigating to a new Add/Edit Gateway page to
prevent stale values.

In addition, I fixed a minor issue where selecting "Add new SBC" and
returning did not have the newly created SBC selected.

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14328 ab1d8caa-1f67-47f1-9e81-24633a41865c