Digit Timer value of "0" wasn't timing out instantly as expected.
Change sipXivr so that a digit timer of 0 or less is considered an
instant timeout.

Fixes XECS-2002 - Call failed to get transfered through "Dial by name"

The "dial by name" section is the only one which collects a single digit,
and the interdigittimeout was set to 0 (meaning "none").  Apparently the
"play silence for 0 duration" command which is used to handle these
timeouts doesn't trigger the "end" event that was expected.  Changed the
code to treat <=0 as 1 mS timeouts.  Also changed the code to not start
a timeout if the timeout value is <= 0.

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14333 ab1d8caa-1f67-47f1-9e81-24633a41865c