XCF-3271: Added indicator icon to show when a conference is locked.
Also added an error message if the user attempts to lock or unlock a
conference that is not active - FreeSWITCH only "recognizes" active
conference, and thus only those can be locked/unlocked. It makes little
sense to lock an empty conference, anyhow, and a locked conference will
become unlocked once it is no longer active.

It is currently not very efficient.

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14766 ab1d8caa-1f67-47f1-9e81-24633a41865c