XX-6027 : sipxbridge: Exception when forwarding a call from AutoAtttendant to user with call forwarding and MOH on bridge enabled.
Tighten up exception handling and do not kill call when MOH dialog
cannot be established.  Bridge handling of MOH leads to race conditions
because the bridge is trying to insert an interaction with the MOH server
concurrently while the call is being transferred. Transfer from AA to
itself or to an extension that immediately forwards the call leads to
this race condition.

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/branches/4.0@15853 ab1d8caa-1f67-47f1-9e81-24633a41865c