sipXtackLib:  Shutdown SipUserAgent properly in SipOutputProcessorTest
Dunno why today of all days the SipOutputProcessTest starting 
dropping core every time (it used to happen occasionally, never twice
in a row where I could catch it).  One of those lovely "pure virtual"
crashes which usually means someone is using an C++ object as an
auto variable and the "random" order of destructors calls ends up
with something being destroyed while some other object is still using
it.

But for some reason, today it decided to die every time and I could
catch a good core dump.  Seems that our old friend the SipUserAgent
doesn't like the destructor called before being told to shutdown.  
We've been through this before, and the quickest solution seems to be
"Don't Do That!".  So I changed the test to call shutdown() before
the auto variable that was SipUserAgent went out of scope and got
destroyed.

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@15932 ab1d8caa-1f67-47f1-9e81-24633a41865c