XX-5943 use new grid component on user phonebook page
New phonebook page table allows for searching and sorting. It adds panel
to present Contact Information.

This is still a prototype stage at the moment. The phonebook grid does
not support clicking on the number, has to be localized. Spacing and
layout need to be adjusted.

Patch from: Ernest Rott - thanks!

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@16345 ab1d8caa-1f67-47f1-9e81-24633a41865c