XX-6637 Registartions page was showing no registrations, when more than 2500 users were registered
Modified building of XML Response body to preset the message body string capacity for lists based on the
number of entries in the list x 650 bytes.  The 650 bytes was estimated from the probable length of
a Registration IMDB record.
Not an ideal solution but should work well for all cases.

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/branches/4.0@16603 ab1d8caa-1f67-47f1-9e81-24633a41865c