EDE http://sipx-wiki.calivia.com/index.php/Express_Development_Environment_Setup
Added 'kraken' to the list of required dependency RPMs since it is now required
for building the sipxopenfire RPM (as of revision 16568.)

Also explicitly add kraken to the list of RPMs when locally building the
dependencies.  Unlike other RPMs (but like sipx-openfire), it is not included
in the 'all' lib/Makefile target.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@16830 ab1d8caa-1f67-47f1-9e81-24633a41865c