XX-7406 SMC : Removal of secondary line failed to update in SMC.
changed the velocity template for SMC phone in order to disable all other lines (up to max lines permited for SMC) except the existing one. For example, if we are having 2 lines and the max is 10 lines, there will be generated all the 10 lines, but the lines from 2 to 9 will be disabled.

Patch from Bogdan Brezoi, thanks!

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@17769 ab1d8caa-1f67-47f1-9e81-24633a41865c