XX-7396 added a small javascript hack for select all checkbox that works in dinamically constructed tables
this patch will assign a new javascript function to the checkbox onclick event.
the function will iterate through all checkboxes in the table and set them true or false.
the script must be included *inside* the refreshForm, otherwise at refresh, the old event will be assigned to the checkbox
This patch also sets the id of the control checkbox to "controlCheckbox" instead of any, which caused some issues if another
"Any" tag was present in the page.

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@17900 ab1d8caa-1f67-47f1-9e81-24633a41865c