XX-7659 The common Branch address is not porpogated to the User or the Phonebook address
Add address table, and replace the address colomns in address_book_entry and branch with pointers to entries in the address table
Add m_branchAddress attribute to AddressBookEntry. This points to the branch address associated with this AddressBookEntry.
Modify the officeAddress getter to account for the branchAddress when retrieving the office address
Add in AddressBookEntry db tests

Based on patch from: Cristi Starasciuc - Thanks!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@18354 ab1d8caa-1f67-47f1-9e81-24633a41865c