XX-7945: Backup and Restore : Fails to restore IP address used as domain alias
Custom validator for the domain alias which must follow this rule: the domain name and the alias should not be the same; however, the domain can be a FQDN and the alias the IP of the FQDN - when initialising the domain do not set the domain name as an alias - put localized validation meesages in sipxconfig-web.properties

Note that
* if we have multiple registered locations, when they are initialized the FQDN and IP addresses
of each of the location will be written in the DB as aliases.
* even if we have the FQDN as the domain name and the IP as alias, and we delete the IP, if we restore the backup
the IP will be set again as alias. THis is due to fact that the list of aliases need to include _all_ the IP addresses of _all_ hosts.
(http://track.sipfoundry.org/browse/XX-2054)

Small changes on top of the patch: error message is added to ManageDomain.properties instead of sipxconfig-web.properties

Patch from: Cristian Starasciuc - Thanks!

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@18473 ab1d8caa-1f67-47f1-9e81-24633a41865c