XX-9819 - some services do not validate aliases (registrar, presence, auth code) - fixed validation - also improved a little validation during services saving
* problem was due to the fact that getBeanIdsOfObjectsWithAlias always returned the bean itself due to a mistake
in the code. If the bean is among the entities with the alias, the validation will pass. The bean was returned
because it was loaded from the Hiber session which was already modified via config UI with the alias value that was
checked. The solution was to load the bean from the DB.