Sending an INVITE through the proxy with an explicit uri to port 5090 will spiral in sipXbridge and deplete media ports
When an itsp account could not be found for the current request a failsafe itsp account was created. An extra check was
added so that this failsafe account will not be created when the outbound ipaddress:port from the request are the same with
the sipxbridge's local or external ipaddress and port. A null itsp account will be returned.

Also added checks to respond with "482 Loop Detected" when an itsp account cannot be found at all.