PT39929099 - In sipXecs 4.6 during RLS shutdown, after terminating dialog subscriptions to phones, RLS resubscribes to phones.
Added a flag, which is disabled by default, to control whether dialog subscribes are sent directly to uri bypassing the
usage of contact information from reg subscribes. This behavior is useful in case RLS communicates with a registrar which
does not handle reg event subscribes. It was disabled because it's implementation is incomplete and resulted in dialog subscriptions
racing with reg subscription during RLS startup (the dialog subscription was deleted and regenerated when the first reg notify was
received) and when RLS shutdowned the dialog subscription was terminated and regenerated again.

NOTE: In case the behavior is needed it can be re-enabled by setting the newly added flag.