PT40350339: Connection rate limiting
- added new tab System > Firewall > Call Rate Limit
- limit rules defined per ip range or per individual IP / subnet, rates defined at sip level packet type or for all sip packets
- rules are applied in order, if src ip matches rule then the other rules won't get processed. Call Rate Limit tab contains buttons to move up down rules
- each rule adds an iptables chain to process incoming SIP packets
- new option for logging call rate limit packets, logged packets in firewall-ratedrop.log file
NOTE: if you set a rate limit for INVITE messages to 2/minute and the want to change it to 4/sec you'll have to delete the rule and recreate it. Otherwise an service iptables restart needed to pick up the new limit.
If you want to limit calls e.g. 2 calls / minute mind that a call involves 2 INVITEs so the filter should be rate of 4 INVITEs / minute interval