PT43856109 - Random crashes of sipXhomer in sqa client code
There was a race condition in StateQueueClient constructor between the startup of keepAliveLoop
thread and initialization of class members. Sometimes the thread would start before the constructor
initialized all members and would crash because it was using uninitialized data. To fix it the
thread should be started in constructor after all members were initialized in constructor init list.