PT46534249 - Callpark - The call is not connecting to call park extension
Collateral bug because of commit 917ebe4d0be3455c97888805ea9397fa25e49156.
Because the initialization of sipXpark was split in two code blocks some
variables which were allocated on the stack in the first block were going
out of scope in subsequent code blocks making the functionality unstable.

This broken behavior was fixed by rejoining the code blocks in one large block.