UC-1089 SQA crashes on startup when DBs not ready
To fix the crash in case MongoDB is not ready the mongo
oplog hooks were removed from the agent. This functionality was
not used after all.

To fix the crash in case RedisBD is not ready the internal
Redis client was disabled. This client was also unused for the
time being because currently there are no SQA clients (publishers
or watchers) which need the persistance feature of the agent.
This will have to be reenabled in case we decide to go
with Redis as the db for SQA.