UC-1166 Improper Call Routing when transffering PSTN call on system configured with multiple branches and gateways at a  branch
NOTE: Backport from release-4.6

For the call transfer to work properly the transferee should be associated with the
same gateway as the transfer initiator. A new custom SIP header called
'X-Sipx-Location-Info' was added to carry out transfer initiator location information.
This header will be added by the proxy in the refer-to header of the REFER message,
and will contain the location of the REFER originator. Later on when the registrar
will use the location information from this header to look for the rigth gateway.
The header will be removed by the proxy when the INVITE message leaves the proxy.