UC-1382 Consultative transfer must point to the exact gateway that accepted the initial INVITE
Added a new proxy plugin which takes care of providing the correct gateway destination for
transferring scenarios which use multiple gateways. The plugin detects calls which have
a gateway as destination and saves the gateway's lineid in top Via header to be retrieved
in the call response. If a call which had a gateway as destination is successful, the plugin
adds a new record in GatewayDestDB with information about the call and the gateway id.
If the call originator is involved in a call transfer scenario the plugin makes sure that
any INVITE request with Replaces would follow the same gateway