UC-1500 bad Record-Route URI added to in-dialog request
Proxy does not add Record-Route headers to mid-dialog requests because it doesn't need to, so,
the CallDestination plugin must not add a Record-Route header in a request when
there is no record route and there is a call-dest in the r-uri. This does not affect
functionality of CallDestination plugin & SipXProxyCseObserver.