Revert "PT-40389463: Support inbound aliases of users and domains processed in one pass."
This reverts commit d025a229279727e1117df36d0239f98822e49222.

Revert "PT-37419533: Call intro conferences fails with 404 / Not Found"

This reverts commit f9c26b7d07384753043385f05def5887732bebd0.

Revert "PT-38071901:  A regression has been introduced after the most recent patches to the issue.   Fetching aliases fails if we delay rewriting of the request-uri host until an alais is matched."

This reverts commit bddca93b76c7e8474b48b3b004526084e56d262f.

Revert "If request-uri is towards an alias domain and a user alias at the same time, normalize the request-uri and do not add a contact.  If user is registered, regdb redirector will insert the final contact."

This reverts commit 9f44bd8eb22621ba48187dba989f1da7b61a0d4f.

Revert "XX-10428 - Alias Redirector does not match rule if domain is also a domain alias"

This reverts commit 703981c31929c312457872c1e7b0569160fdf511.

Revert "XX-10428 - Alias Redirector must be evaluate prior to regDB redirector so that the correct identity is used to search."

This reverts commit b5026c54caf436132bbe5084628cb89c11f0cf63.